+++ compose/search/src/main/java/com/hegunhee/compose/search/navigation/SearchNavGraph.kt
@@ -1,25 +1,24 @@
-package com.hegunhee.compose.search
+package com.hegunhee.compose.search.navigation
 
 import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.runtime.CompositionLocal
 import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.compositionLocalOf
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
+import com.hegunhee.compose.search.SearchScreenRoot
+
+const val SearchTwitch = "search-twitch-route"
 
 val LocalPaddingValues = compositionLocalOf { PaddingValues(0.dp) }
+
 fun NavGraphBuilder.searchNavGraph(
     paddingValues : PaddingValues,
     onNavigateTwitchChannelClick : (String) -> Unit
 ) {
-    composable(route = SearchNavGraph.searchRoute) {
+    composable(route = SearchTwitch) {
         CompositionLocalProvider(LocalPaddingValues provides paddingValues) {
             SearchScreenRoot(onNavigateTwitchChannelClick = onNavigateTwitchChannelClick)
         }
     }
-}
-
-object SearchNavGraph{
-    const val searchRoute = "search"
 }
